Allow "content" and "project_id" to be permitted
In ruby the last line of a function is the value that will be returned.
Previously we had two separate lines of code for each permitted parameter,
but only the last one was ever used. This change fixes the "tracking_params"
function to permit all the parameters on one line.